requirements.txt
qdrant-client>=1.6.0
docling>=0.6.0
sentence-transformers>=2.2.2
# For chunking and RAG processing
langchain>=0.0.267
# For document conversion
PyPDF2>=3.0.0
markdown>=3.4.3
beautifulsoup4>=4.12.2
# For visualization (if needed)
matplotlib>=3.7.1
plotly>=5.15.0
# For data handling
pandas>=2.0.0
numpy>=1.24.0
# For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
streamlit>=1.26.0
langchain_openai>=0.0.1
langchain_community>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0